Move captured_stdio() to tests.helpers._io.
